write onnx engine

+++ include/json/jsonbuilder.h
@@ -1,5 +1,5 @@
 #include <string>
 #include <map>
 #include <variant>
-typedef std::map<std::string, std::variant<float, int, double, std::string>> vmaps;
+typedef std::map<std::string, std::variant<int, float, double, std::string>> vmaps;
 std::string create_json(const vmaps &kvs);
